Update trabalho av2(1).py

+++ trabalho av2(1).py
@@ -1,5 +1,5 @@
-base_maior = int(input('Base maior: '))
-base_menor = int(input('Base menor: '))
-altura = int(input('Altura: '))
+base_maior = float(input('Base maior: '))
+base_menor = float(input('Base menor: '))
+altura = float(input('Altura: '))
 area = ((base_maior + base_menor) * altura) / 2
 print(f'Esse é o resulatdo: {area}')
